requirements.txt
Flask
BeautifulSoup
python-nmap
